use useUniqueTimeSelection composable for timestamp management
